created modular flag for input actions. changed implementation of attack to sync over network
modular flag for input actions, when flagged on, binds that action to  shift/alt/ctrl + that key automatically to allow the function to process any of those keys itself.

Changed attack to send a damage signal over the network; only the server's spawned reticles will actually send the damage signal.